Key fixes for INI handling; started work on edit gizmo color schemes
Fixed bugs with how we load and save main.ini. The `std::ifstream` class closes whatever handle you give it. We were trying to use QFile and `std::ifstream` together, but that meant we'd end up closing the handle twice (once from the STL and once from QFile). In general Qt's utility classes are not interoperable with the STL, so you really do *have* to pick one or the other; in this case, the STL wins out.

The options subsystem is now loaded on a single-shot timer, so I don't have to worry about potential deadlocks should a DovahkitCore-dependent subsystem try to access the options subsystem (directly or indirectly).

Added an `options_collection` class that can be subclassed; this can be used to have groups of specialized options register themselves with the options system for some basic central coordination. (You could also use signals and slots, but those are more freeform; the system I've done here implies a manager/subordinate relationship.)

Started work on edit gizmo color schemes. Currently, we have a list of hardcoded schemes, and basic connections to get those working.

Added icons for use in the UI. Started work on the UI for edit gizmo color schemes.