ugh
Work on the UI for changing edit gizmo color schemes. We now use a model, in order to more easily be able to open an edit dialog. That model is powered by a new DKGenericListModel, similar to DKGenericTreeModel.

Very little of this has been tested because MSBuild is too busy being a terrible build system from hell