Started work on Worldedit tool refactor
Previously, Worldedit would invoke all tools behaviors manually within its input polling function (i.e. it'd call Worldinput to get requested actions and then immediately perform those actions itself). This is not going to be maintainable going forward, so what I'm doing now is extricating the code for each tool out of the Worldedit core and moving it into the tools themselves. This is what the previous commit was meant to facilitate, by both clarifying terminology and making available the term I wanted to use for this: "invoking" a tool.

There's a design challenge I need to resolve before we can do this for camera tools, and a passkeyed accessor I need to add for the landscape debug tool. I have at least moved tool invocation code for most tools into the tools themselves; the goal is to eventually be able to do a compile-time loop over a cobb::class_array within Worldedit, so I don't even have to manually write the if-statements and calls for each tool.

Some of the accessors I've added to the Worldedit core, for tools' sake, are public; this is because we may want to be able to activate these same functions via the UI.