Started work on a `move_selection` tool
Started work on implementing functionality in Worldedit to move and rotate the selected refs. Currently, movement seems to work(-ish), but we don't update the 3D render after modifying the REFR data.

We currently have a test button on the Render Window toolbar. When using this, be sure to angle your view so that only empty void/landscape is behind the button: due to an unrelated Worldinput bug, clicks on the Render Window toolbar are detected by the 3D view and can select things "underneath" the toolbar.