Worldedit: selection move/rotate work
The Vulkan-side code now offers a helper function for setting the root transform of a rendered NIF.

We now update 3D for a REFR when we move or rotate it.

We now update the rendered bounds for a REFR when we move or rotate it.

The logic for updating a selected REFR's coordinates is a bit more involved now, but less prone to floating-point imprecision. We only use rotation matrices and decomposition for rotation about a centroid when multiple refs are selected; translation is performed separately afterward, and for a single ref we just modify the rotation directly.

The test code in the Render Window now converts rotations from degrees to radians before passing them to Worldedit.