Worldinput fixes
Fixed a bug in the "Debug WASD" control scheme where we were accidentally setting a non-single-control ISG as the IS's raycast-associated (not-a-)button, causing assertion failures in the control scheme editor. I should probably have the hardcoded control schemes double-check and assert validity on themselves when in Debug.

The options window now properly allows you to delete control schemes; the underlying model had a mistake in its `removeRows` implementation.

The Render Window and Worldinput no longer react to mouse clicks outside of the 3D view, i.e. clicks on the toolbar or status bar no longer count as clicks on whatever 3D objects are "beneath" those bars.