Worldinput config constants are now run-time-editable
The following configuration values in Worldinput are now implemented as user-alterable INI settings rather than `constexpr` constants:

Press to Long Press threshold: the amount of time in seconds for which a key must be held down, for its release to count as a Long Press rather than a Press.

Press to Hold threshold: given two control scheme actions that map the same keys but with Press versus Hold, the amount of time for which the buttons must be held in order for the Hold action to "win." For a Long Press/Hold conflict, the Press to Long Press threshold is added to this threshold.

Key sequence expire time: For ordered-sequential input sequence groups, the maximum amount of time that can pass between the release of one child ISG and the press of the next; if more time passes, the user fails the button combo.