Current WI scheme now tracked by filename, not user-facing name
We now track the current control scheme by its filename, not its user-facing name. This means that if ever we fail to rename a control scheme to use a unique name (which is harder than it sounds), it doesn't really matter, because we'll still generate a unique filename for it.

Fixed major bugs in my INI helpers, which caused us to re-save INI content incorrectly.