Critical Worldedit fix; other minor fixes; docs and future planning
Critical fix to Worldedit: we now react to `dataAbandonImminent` and unload everything, which ensures that if you switch from one load order to another in a single program session, we don't hang onto `form_stub` pointers and cause problems.

When Worldedit unloads all cells, it now properly unloads REFRs that were not in a loaded cell. (Remember: we made that possible with moving selections outside of the loaded area.)

When Worldedit is directed to load nowhere (a null cell/world pointer), it now resets to the initial render state more fully: in addition to showing the grid as it did previously, it now resets the lighting and camera coordinates to their defaults, too.

Fixed improper buffer management in the Render Window's world axis helper overlay: we were mishandling buffers when updating the overlay in response to gizmo color changes. Just some sloppiness when we initially added support for gizmo recoloring to that overlay.