Factored out Vulkan camera; work on an `orbit_camera` tool
Split the camera-related functionality in the renderer into its own class, rather than having `scene` own those calculations. This has made the camera code more manageable and more cleanly separated.

Fixed the implementation for strict ordering of Worldedit tool results.

Started work on an `orbit_camera` tool. The tool itself works, but the underlying functionality it calls into (`vulkanDK::camera::arcball`) is broken. The Reach control scheme has been updated to use the tool.

---

The problem with the "orbit camera" tool is, bluntly, that there are zero good resources on the web for this whatsoever. There are tons of resources for implementing an FPS camera, and there are tons of resources for implementing an arcball camera, but every single one of these resources is flawed. The vast majority of sources assume that your camera type will never change -- that your camera will never *become* an arcball camera or *stop* being an arcball camera -- so many of them take the shortcut of only pivoting around the world origin, building this fundamentally into their view matrix computations. The vast majority of sources also fail to clearly define a consistent convention for their axes; even the ones that bother to specify their Euler order fail to specify their handedness, for example. Many sources are just straight-up wrong, though this is provable only by testing them against virtually *every* convention.

The online sources that come closest to working are designed to support both camera modes at once, but with too much extra state: they arcball by default, but allow the camera to move and rotate independently as an additional modifier on top of the arcball movement; and they manage this by storing arcball state *and* FPS camera state together. It's grossly unnecessary! If you have the camera's current position and orientation, the pivot position, and the angles by which the camera's orbit should change on the current frame, then that's all the state you should need.

As soon as my Internet comes back online (it's down as of this writing), I plan to stop searching for anything camera-related and anything arcball-related, and just search for the general case of making one object orbit around another.

The current status of the arcball camera is as follows: if you select an object, you *can* successfully orbit around it... roughly. If the object wasn't centered in the view at the time it was selected, then your orbit will be off-center. To fix this, we'd have to force the camera to aim at the target object (at minimum when we start arcballing, but if we do it *correctly* then we can just do it on every frame for which we're arcballing), and that is what I'm currently struggling to implement. This is not helped by how fuzzy the Euler conventions within the program are (I know Skyrim's conventions, but GLM's conventions vary depending on what macros are in effect at each point its headers are included, and on whether any given library function bothers to *check* those macros, which isn't done consistently, and on the phase of the moon...) and how fuzzy conversions between them are as a result.