WI: range control conflicts; WE: invocation-in-tandem docs
Documented the system for designing Worldedit tools that are "invoked in tandem."

Worldinput: binds with range constraints can now conflict with one another, if they consume overlapping axes on the same range control. A conflict resolution algorithm has been implemented and added to the spec to cover this. With this change, the Reach control scheme's zoom bind (LS Move to move a selection; RT + LS Move Vertical to zoom) almost works as intended. (The problem is that deadzones don't get reapplied to the X-axis-constrained "move" bind, so you get slight movement alongside the zooming.)