Vulkan: inverted view/proj matrix optimizations
It turns out that the issues seen with drag-moves in testing were not actually a precision issue (although precision issues *did* cause us to drop a pixel or two); rather, the biggest issue stemmed from a logistical problem within Worldinput. The bind for selection (`Press LMB`) was conflicting with the bind for dragging (`Hold LMB :: Mouse Move`), causing the latter to be delayed per Press-delays-Hold conflict resolution. I'm considering options for fixing this, given that the Hold bind in this case has a range constraint and that should count for something.

Still, the renderer now has functions for computing inverted view and projection matrices much more optimally. The `surface_position_to_world_ray` function now uses the same math, though I've thought of another way to optimize it...