Worldinput: new condition for current gizmo mode
Worldinput: added a new condition for control schemes: the current edit gizmo mode.

Updated the Debug WASD control scheme: Numpad 7, 8, and 9 will apply a positive transform on the X, Y, and Z axes; Numpad 4, 5, and 6 apply a negative transform. The type of the transform matches the current edit gizmo mode (i.e. if the translate gizmo is showing, these binds translate the selection). Note that rotation and scale are not yet implemented, as tools for those tasks are also not yet implemented.