Fix: range-only sequences now stringify properly
Fixed a bug in input sequence stringification that caused failures when trying to stringify range-only input sequences. This means that these sequences will now display properly in the control scheme editor UI, since it uses the same algorithm.

Fixed a bad assertion in `input_sequence::assert_validity` that would've broken range-only sequences.

The Worldinput spec now leads with an explanation of what Worldinput is and does.