Started work on ACTI UI and backend Papyrus rewrite
Started work on a UI for editing Activators. Realized I needed a UI for editing attached Papyrus scripts. Realized that requires backend rewrites for how we handle Papyrus.

Realized we were only handling "raw" Papyrus data. When a Papyrus script is attached to a base form, REFRs can selectively override some or all of the script (including selectively removing it i.e. opting themselves out of having that script), and we had no constructs or code to handle this.

As such, we've reorganized all of the backend Papyrus structs, and are currently working on code to get the "computed" ScriptObjects for any given form in a manner that is relatively easy to edit and track. The idea is, you'll "check out" the Papyrus data for a form -- so we resolve base/ref relationships within the data -- and make your changes, and then "check in" that data -- at which point we figure out which of the scripts you've listed exist on the base form versus which don't, etc., etc., and edit the target form's Papyrus data. Once this code is ready, we can integrate it into the WIP Papyrus UI (tearing down much of what we've already built for that, lol). When we have that able to compile, we can build the rest of the ACTI UI and then begin testing.

The Lua APIs for Papyrus are currently commented out and will need to be replaced. The plan is to have them work the same way as at the native layer (once that's working), i.e. the getter for Papyrus data returns a copy (as a userdata, not an arbitrary table, so we can still enforce typing) that you can modify without any live-updates, and you then assign that back in once you've made your changes.