Fixes
The `property_status` enum now declares its underlying type, fixing a serialization mistake.

Various compiler errors fixed; the program now builds again.

The widget and model for the list of Papyrus scripts on a form now operates on a working copy rather than on a form stub. This is necessary in order to, uh, *use* these on working copies at all. The model for the list of properties on a script isn't updated yet, but I have... plans... for that.

To-do list updates.