Work on Papyrus UI
Phasing out the separated models for VMAD script and property lists, in favor of a single model that'll manage the entirety of a working copy's VMAD. The properties dialog will just scope its tableview to a different model node.

Continued work on implementing that properties dialog. We can now display scripts and their properties. One caveat: we'll need something like a QIdentityProxyModel subclass for the properties dialog, in order to be able to edit property values (and display edited values) yet only commit changes if the user clicks "OK."