further down the rabbit hole we go
VMAD model: started work on the machinery for editing properties' values.

Started work on a widget, intended to be usable in Qt Designer, for picking a REFR (e.g. for a condition argument, a Papyrus ObjectReference property value, etc.), because nested problems are great and life definitely isn't pain and suffering! :D :D :D

DKRefsInCellModel was designed in the hopes of avoiding the need for QSortFilterProxyModel and is self-sorting; however, implementing filtering would be... difficult. I'll give it a try and if it doesn't pan out, I'll simplify the model in favor of wrapping it in a QSortFilterProxyModel.