FormPicker fix; DKObjectReferencePicker progress
Implemented the button to show the current ref in the Render Window. Picking is still not implemented.

PlayerRef is now always made available as a selectable option.

Fixed sorting issues within DKRefsInCellModel.

Fixed mistakes made when implementing filtering.

Fixed mistakes made with the last commit's changes to FormPicker.

At this point, I think the only unfinished feature in DKObjectReferencePicker is being able to pick a ref via the Render Window. All that's left now is to get it working and placeable in Qt Designer.