the program now builds again
Created a new widget for selecting quest aliases: DKQuestAliasPicker.

Fixed focus and tab order for DKObjectReferencePicker.

DKScriptObjectDialog now has a DKQuestAliasPicker.

Did some tests in the CK and in-game and confirmed that mixed arrays of forms and aliases are not viable in Skyrim.

Fixed compile errors related to the BSA changes in a previous commit.

Added a new type, `ui::types::quest_alias`, for allowing the UI to more easily pass around quest aliases (as a quest-and-alias-ID pair). This type is also QVariant-enabled (courtesy of DovahKitCore's constructor).

Started work on trying to update DKFormVMADModel to only allow arrays of forms or arrays of aliases, but not mixed-representation arrays; this would make it easier to code the UI for editing property values. Put this work on hold so I can try to test the underlying Papyrus model.

Started work on extricating some Papyrus-related UI data types from DKFormVMADModel into namespace `ui::types::papyrus`. This, too, is a WIP. At this point, trying to add Papyrus editing to DovahKit has resulted in me having six or eight ongoing WIPs at the same time, nearly all of which depend on each other in some way. I hate being alive

---

Current goal is to test creating a DKPapyrusModel and see if it properly grabs all PEXs across all BSAs as well as loose files. If so, we can start work on having DKScriptObjectDialog validate incoming VMAD data, using DKPapyrusModel to map PEX-side property typenames to underlying native types. Once that's done, we'll want to implement the change listener for loose files and test that; once THAT'S done, we'll then want to move on to having DKPapyrusModel track what forms list what scripts in their VMADs (which in turn is needed for properly filtering form/ref/alias pickers in the property editing UI).

That said, there are things in DKPapyrusModel that we should move to the backend: specifically, the map of scriptnames to native types, and information about which native types subclass each other. The latter is something we want for `ui::types::papyrus::vmad_property_value_to_ui`.