PEX refactoring
Wrote new types and a new parser for PEX file structures. The goal is to be able to support multiple readers -- most pertinently, a reader that skips absolutely everything except the game ID in the file header, and the names, superclasses, flags, and docstrings of all scripts in the PEX data.