PEX parsing: work on a class-info-only parser
Created cobb::interned_string_table as a configurable implementation of string interning.

Modified the new PEX parser so that 99% of the file-reading code is in a CRTP base class. This makes it easier to create alternate parsers.

Created a new PEX parser which skips everything unrelated to a class's info and hierarchy. It's meant to use an interned string table to store that data, but I, uh, forgot.