Benchmarks; prep for known-scripts refactor
Ran benchmarks for five specific PEX files, with 50 tests per parser. Full parser takes 358ms; skim parser takes 5.7ms, or 1.6% the amount of time.

If I modify DKPapyrusModel to use the skim parser for scanning PEXs, I expect a reduction from a predicted 200s load time to 3s. I'd still like to make PEX indexing async post-load, which is one of multiple reasons why DKPapyrusModel needs a refactor, but that'd be a massive savings; intend to test that in next commit.

---

Speaking of which, the refactor:

I want to use a Qt model only as a facade, and have known scripts (i.e. `DKPapyrusModel::Script`) be managed by a dedicated subsystem which offers only the bare-minimum Qt signals needed for a model to keep up. In turn, `DovahKitFormDataCache` will associate each VMAD-capable form stub [that has any scripts attached] with a list of refcounting pointers to known scripts, so we can properly track when scripts are known and when they cease to be known. To that end, started work on refactor; defined `known_script` type though I haven't moved the bulk of DKPapyrusModel code out yet.