DKPapyrusModel take 2: failed
It now takes 1:47 for DKPapyrusModel to index all PEXs -- significantly faster than before, but still bad. It grows slower as the process grows longer, and this is no doubt due to how we store the known scripts.

I created a purpose-built list type for listing Papyrus scripts sorted by name: it counts the number of entries that start with A, the number that start with B, and so on, so it can skip huge portions of the list based on the first letter in the scriptname to look up. In practice, though, this is not terribly helpful due to fragment scripts: there are 5525 files that begin with `TIF_` in the base game alone. Plus, the underlying storage is a vector (bad habit) so when we insert into the middle of the list, we have to shift everything, which incurs massive overhead as well.

Gonna try splitting the storage into twenty-seven QLists (a twist on the optimizations currently in place for alphabetic lookups) but I suspect I'll need another approach entirely (e.g. unordered map internally; then generate a sorted list of scriptnames post-indexing).