Fixes to handling loose PEX changes
FIX: When a loose PEX was created or modified to subclass a native class, we accidentally created "phantom" scripts for those native types because we inadvertently `continue`d a nested loop when we meant to continue the outer loop.

FIX: When a loose PEX was modified, we did not consistently clear its `extends` info (superclass pointer, underlying type, etc.).

FIX: When a loose PEX was modified to subclass another, already known, script, we failed to establish that relationship properly.