Work on VMAD scanning: seems to be working!
Created `known_script_ptr`.

Built a replacement for DovahKitFormDataCache: the `form_info_cache` subsystem. This subsystem is designed to grab model paths for base forms of interest, quest filter strings, and info for all attached scripts on all forms. Currently, the program doesn't actually instantiate it directly, but there's a debug menu item to do so: you can test it by invoking that debug menu item and then loading game data.

In tests, it takes about a full second for the form-info-cache subsystem to skim record data for all major base forms' model paths, for all quests' filter strings, and for attached script lists on all VMAD-capable forms [that we currently have code to load in full]!

---

The refcounts on known scripts are now atomic, to account for us creating refcounting pointers on multiple threads during the form-info-cache process. As a result, some member functions on known scripts can no longer be constexpr, because std::atomic isn't really constexpr-capable.

Disabled updating of scriptnames' letter case when scanning a changed PEX, to work around limits in the CK and Papyrus Compiler.

The Papyrus subsystem keeps track of whether it's currently in teardown, to optimize destroying refcounting known-script pointers en masse as a result of teardown.