Form-info-cache now used by Object Window; old cache now unused
Cleaned up some of the metaprogramming used in the form-info-cache.

Fixed bugs that prevented the form-info-cache from properly caching attached Papyrus scripts. With these fixes, it now takes a full second to cache everything in Debug, which is pretty impressive given the sheer bulk of the data it has to work with.

The Object Window's treeview and filtering system now use the form-info-cache instead of DovahKitFormDataCache. Additionally, the filtering code has been cleaned up, relying on the fact that the treeview doesn't allow multi-selection.