Form info cache now handles quest alias scripts (untested)
The form info cache now gathers scripts attached to quest aliases and makes them separably queryable from scripts attached to the quest itself.

The Papyrus subsystem now allows you to query whether a form has a given script (or any subclass) attached, accounting for REFRs inheriting scripts from their base forms.

Fixed a bug in the QUST loader that would, theoretically, lead to incorrect parsing of alias-related VMAD content if that content were edited to contain data that is valid for the game but impossible to generate in the CK (and a logistical nightmare to support in DovahKit, which is why we don't).