DKAllPapyrusScriptsModel and DKAddPapyrusScriptModel first draft
Built DKAllPapyrusScriptsModel as a singleton, and DKAddPapyrusScriptModel to power the "Add script" we want to have when editing a form's VMAD.

The Papyrus subsystem now makes it possible to loop over all known scripts, to facilitate the above.

The Papyrus subsystem now makes it possible to query whether the initial known script discovery process (PEX indexing within the Papyrus subsystem + VMAD scanning by the form-info-cache subsystem) is complete.