Papyrus: "Add script" dialog now exists and is functional
Can only test it on ACTI forms for now, so we can't test the "Show Hidden Scripts" checkbox (no REFR-attachable scripts are hidden; what would be the point?). Still, we know that filtering by text works, and limiting the list to scripts attachable to the current form (including not allowing scripts already listed on the form) works.