`dovah::form_type` is now a strict enum + misc fixes
Turned `dovah::form_type` into a strict enum. This necessitated changes all over the program. This has benefits of its own, but my main hope is that it'll make it more straightforward to make the enum Qt-enabled, for Qt Designer widgets that will need it (i.e. DKFormPicker).

---

The added type-safety of using a strict enum made it possible to find and fix several places in the backend where we looked up form type info by index even when the form type list diverged from the form type enum. (That is: given a form type `ft`, past a certain index, `dovah::form_types[(size_t)ft]` is *not* the info for `ft`, and we were mistakenly assuming it was.)

Conditions on topics and topic-infos now properly retrieve the parent quest.

Dovahscript: although topic-infos should not be able to exist outside of topics, we now null-check the parent topic when getting the parent quest.

Dovahscript: REFR wrapper: simplified the code for managing what REFR form flags are available for what base form types.

Band-aid-fixed an edge-case that could cause an assertion failure in the Vulkan renderer during program shutdown: if we shut down with the right timing such that some scene entities [that own GPU resources] failed to properly sync their lifetimes, then we'd fail an assertion meant to guard against leaked GPU resources.