Started work on DKBoundScript(List)Model
It turns out: it actually *is* much simpler to use separate models for the list of scripts on a form, versus each individual script. The trick is to have the former model own the instances of the latter model.

Implemented these separate models, intended as replacemnets for DKFormVMADModel, in the form of DKBoundScriptListModel and DKBoundScriptModel. They aren't hooked up to anything yet.

Removed the "Add Property" button from the script-object dialog: this button, from the original CK UI, is used to actually edit the script i.e. PSC+PEX itself, not the BoundScript. We don't want that.

The class for VMAD data now exposes a compile-time constant indicating the maximum number of scripts that can be attached, as determined by the file format.