DKBoundScriptModel no longer displays hidden properties
Properties that are flagged as "Hidden" in a PEX will no longer be visible in the UI via DKBoundScriptModel. The model still retains the properties, and will retain any values that have been set on the properties (e.g. if a value was set and then the script was recompiled to hide the property, or if a value was set via xEdit).