RE for TESModelTextureSwap, and changes in response
Did some reverse-engineering of the Creation Kit and the game to figure out exactly how NIF texture swaps are handled. Turns out, neither one uses NIF block indices. Rather, they use an alternate indexing scheme ("leaf indices"), with the game relying on leaf indices and the CK relying on block names.

Renamed fields on our TESModelTextureSwap analogue and associated data structures to reflect that the indices in question are leaf indices.

Since leaf indices are non-trivial to figure out, DKFormNIFPickerDialog now relies on that old helper function in nifDK (which, of course, has been rewritten to account for these new discoveries).

Rewrote `nifDK::file::apply_texture_swaps`, used by the Render Window, to mimic the game's logic: we use leaf indices rather than block names, now; we create BSShaderTextureSets for any geometry blocks that lack one but are the target of a texture swap; and we also apply the special case of culling out any geometry nodes that are the target of a NullTextureSet texture swap. We prefer leaf indices over block names (contradicting the CK) because if you load an ESP and there's a form whose NIF has been edited, there's no guarantee that DovahKit will end up updating that form's model data (since we lazy-load everything); in fact, if it's a form in a master, there's no guarantee that DovahKit or the CK would override it to update the model data.