Fixed file selection when picking NIFs
Fixed bugs in DKBSABrowseDialog that would cause the list pane to get stuck blank when switching from the initial List View, to Icon View, and then back to List View. We can't use a nullptr delegate, and QSize{} is not the same as QSize(0, 0).

Fixed path handling in DKFormNIFPicker.

Added documentation comments to DKGameFilePicker to clarify its (currently unclear) path conventions. Maybe we can look into revising that post-launch.