DKFormNIFPicker done; started work on destruction stage UI
Qt Designer can now be used to place DKFormNIFPicker widgets.

The Activator dialog now uses a DKFormNIFPicker via Qt Designer.

Moved `DKFormNIFPicker::_state` to a new type, `ui::types::nif_for_form`, since the destruction stage UI needs it too.

Started work on the UI for destruction stage data. The button to open the dialog is ready, but the dialog itself is not. Layouts for both nested dialogs have been built, as has the dialog for individual destruction stages. The dialog for destruction data as a whole still needs to be built.