DKStyledItemDelegate; destruction stage flag checkboxes now centered
Created DKStyledItemDelegate as a catch-all for any enhancements or fixes we need to make to how Qt displays certain things. Currently, it's set up so that if a view item contains a checkbox, no display role (or a default-constructed QString), no decoration, and a text-alignment role, then the text-alignment property applies to the rendered checkbox.

Destruction data UI: flag checkboxes are now centered, using the above.