QLinedTreeView fix for children of a last-child node
Fixed a bug with the dotted lines on the Object Window treeview. Consider the tree nodes described by the following CSS selector:

   *:not(:last-child) > *:last-child > *

These tree nodes should have a dotted line to their immediate left, a blank space past that, and a dotted line past that; that is, the space below the last-child node should be blank. The bug was that we were instead drawing half-tall dotted lines for each row in that space.