Fixes for last commit
Converted `length_prefixed_string_was_too_long_to_save` from a notice code to a new error object, since the machinery needed to support the former in that situation was removed. This necessitated adding `_current_target` to `file_writer` so that we could find the form stub being written when the error needs to be emitted -- a hack; again, I really, really want to rewrite the backend stuff someday.