Update file_load_order.cpp
FIX: If `file_load_order::accept_form_stub` throws while handling an override, then we end up with a nullptr entry in the forms map, because we check for an existing (i.e. to-be-overridden) form using `operator[]`.

We now use `find` to check if an existing (i.e. to-be-overridden) form exists.