Fixed crash on re-throwing `file_save_failed` across threads
Fixed a problem with capturing certain exceptions using `std::current_exception()`. This was causing us to crash when failing to load a file. I don't know if this is a hole in the standard, a problem with MSVC's implementation alone, or what, but I've documented the problem in the Markdown file included in this commit.

The fix shouldn't actually be necessary for `file_save_failed` because we don't multi-thread file output (and therefore don't use a `std::exception_ptr` for it), but I've made the change anyway.

With these changes, tests now confirm that file load failures are handled within the program with no visible mistakes or problems.