Update rewrites I want to do someday post-launch.md
Plans for rewriting file handling.

Plans for rewriting Dovahscript.

Removed the section on notice codes, since we did that early (to avoid the work piling up and getting even worse as development continues). Added a section on some finishing touches for that change that I want to push to post-launch.