GMST warning fixes; log window tweaks; XRMR loader improved
The log window now displays appropriate icons for its own built-in messages (none-stubs; file load complete).

Loading for XRMR and friends now matches the game directly, and emits appropriate warnings.

Fixed how some warnings for bad GMST records are emitted. We no longer emit a warning for the subrecord having "extra data" when the setting type is unidentifiable (since we can't, at that point, know how large the subrecord should be). We now stringify the form ID for the "unknown type" warning properly.

We now enforce the maximum number of linked rooms for XRMR and emit a save error when that constraint is violated.