Form-editing dialog refactor
Completely refactored how form-editing dialogs are implemented. The code isn't and can't be as clean as I want it to be, due to limitations imposed by Qt's build system. However, it's quite a bit cleaner than it was before. Specific improvements include:

 - You no longer need to specify a dialog's form type in two places (once via a macro to hook up the loaded form class; and then again via a manually-passed constructor parameter).

 - Certain checks for errors that should be impossible (i.e. opening a dialog for form type X on a stub with form type Y) have been turned into assertions.

 - Form working copy dialogs now store a stub's editor ID and record flags locally. This means that you can bind controls to these values, i.e. using `ui::bind`; that is, you can have your UI controls update the (local) values in real time, and then the dialog will automatically commit them to your stub when "saving."

 - Dialogs now access the data to be edited in the same manner (`this->form`) regardless of whether it's "real" or a working copy.

---

The functions for opening a form-editing or use-info dialog for a form now take the form as a reference rather than a pointer, as nullptr values don't make sense here and were never checked for here.