Conditions refactor + DKConditionList; partially tested
Fixed missing parameter on IsPlayerInRegion definition.

Attempted to rewrite how conditions are defined and handled throughout both the backend and the frontend.

 - All condition function information is now `constexpr`.
 - All condition function parameter typeinfo is now `constexpr`.
 - Condition function information no longer includes the description, as that text isn't useful for raw code.
 - The data structures now do a very marginally better job at enforcing data integrity via strategic use of std::variant and std::optional. True perfection within the "working condition" struct will never really be possible due to how interconnected and interdependent the fields are.

The condition list widget is now usable in Qt Designer as DKConditionList.

---

The program builds, and the UI can properly display condition lists. The dialog box for editing conditions also properly shows the fields' values. Tests for actually making changes to the data are still pending.