Compiler fixes; critical file-saving fixes
Fixes to Windows-related headers and includes for when compiling as non-Unicode.

Updated the project to specifically compile as Unicode, instead of trusting the compiler defaults which, evidently, can change.

Fixed a mistake made when refactoring/moving the file-saving code, which caused us to try (and fail, thankfully) to delete forms that were not unsaved.

Fixed a mistake made when refactoring/moving the file-saving code, which caused us to fail to save new active files (because the desired filename wasn't being passed from the UI to the save process).

Fixed a mistake in the Papyrus serialization code that caused us to write the status enum as a four-byte value instead of a one-byte value.

Changed some calls to `unchecked_read`, while loading form data, to avoid potential crashes. At some point in the future I'll refactor this dreadful file-reading code...

Removed unused value `parameter_underlying_type::event` within the conditions backend code.